Some cleanup, documentation, refactoring.
