+++ src/main/kotlin/dniel/forwardauth/infrastructure/spring/filters/LoggingUserFilter.kt
@@ -1,7 +1,8 @@
 package dniel.forwardauth.infrastructure.spring.filters
 
+import dniel.forwardauth.application.CommandDispatcher
+import dniel.forwardauth.application.commandhandlers.AuthenticateHandler
 import org.slf4j.MDC
-import org.springframework.core.annotation.Order
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Component
 import javax.servlet.FilterChain
@@ -13,7 +14,8 @@ import javax.servlet.http.HttpServletResponse
  * User details can be found in the Auth0 Dashboard: https://manage.auth0.com/#/users -> Search the "user_id" by Lucene Syntax
  */
 @Component
-class LoggingUserFilter : BaseFilter() {
+class LoggingUserFilter(authenticateHandler: AuthenticateHandler,
+                        commandDispatcher: CommandDispatcher) : BaseFilter(authenticateHandler, commandDispatcher) {
 
     override fun doFilterInternal(req: HttpServletRequest, resp: HttpServletResponse, chain: FilterChain) {
         trace("LoggingUserFilter start")
